requirements.txt
requests>=2.32.0,<3.0.0
dnspython>=2.6.0,<3.0.0
pyinstaller>=6.0.0,<7.0.0
